+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Minimum Supported Rust Version (MSRV): 1.46
+- `EBUSY` added to libseccomp error enum
 
 ### Changed
 - `get_api_sysrawrc` and `set_api_sysrawrc` can now be used with any API level.